Fix: Change model to decimal field.

+++ locations/models.py
@@ -2,5 +2,5 @@ from django.db import models
 
 class Location(models.Model):
     type = models.IntegerField()
-    long = models.FloatField(read_only=True)
-    lat = models.FloatField(read_only=True)
+    long = models.DecimalField(read_only=True)
+    lat = models.DecimalField(read_only=True)
